requirements.txt
httpx==0.28.1
fastmcp==2.10.5
python-dotenv==1.1.0